[.github/workflows] Added pigweed-app related builds to workflows.
There is no build validation in CI for pigweed-app,
as this app related builds are not ran in the github workflows.

* Modified nrfconnect_example.sh to take board name as an mandatory
argument and also allow passing list of additional arguments
(e.g. attaching overlays).
* Added pigweed-app build and lighting-app with rpc.overlay build
to the nrfconnect github workflows.
* Renamed esp_echo_app.sh to the esp_example.sh and modified it
to take application name as an argument.
* Added pigweed-app build to the esp32 github workflow.
* Renamed Echo App build step to the All Clusters App in the esp32
and qemu github workflows, as in fact that app is being build.